Tutorial_1_Docker/ejercicio2/compose/app/requirements.txt
slackclient==2.9.3
pymongo==3.11.2